+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmObjectHelper.kt
@@ -104,7 +104,7 @@ internal object RealmObjectHelper {
     internal inline fun <reified R : Any> getList(
         obj: RealmObjectInternal,
         propertyName: String
-    ): RealmList<Any?> {
+    ): ManagedRealmList<Any?> {
         return getList(obj, propertyName, R::class)
     }
 
@@ -112,7 +112,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectInternal,
         propertyName: String,
         elementType: KClass<R>,
-    ): RealmList<Any?> {
+    ): ManagedRealmList<Any?> {
         return getListByKey(obj, obj.propertyInfoOrThrow(propertyName).key, elementType)
     }
 
@@ -120,7 +120,7 @@ internal object RealmObjectHelper {
         obj: RealmObjectInternal,
         key: io.realm.internal.interop.PropertyKey,
         elementType: KClass<R>,
-    ): RealmList<Any?> {
+    ): ManagedRealmList<Any?> {
         val o = obj.getObjectPointer()
         val listPtr: NativePointer = RealmInterop.realm_get_list(o, key)
         val mediator: Mediator = obj.getMediator()
@@ -140,7 +140,7 @@ internal object RealmObjectHelper {
         clazz: KClass<*>,
         mediator: Mediator,
         realm: RealmReference
-    ): RealmList<R> {
+    ): ManagedRealmList<R> {
         return managedRealmList(
             listPtr,
             ListOperatorMetadata(
@@ -329,8 +329,14 @@ internal object RealmObjectHelper {
         }
     }
 
-    @Suppress("UNUSED_PARAMETER")
-    internal fun setList(obj: RealmObjectInternal, col: String, list: RealmList<Any?>) {
-        TODO()
+    @Suppress("unused") // Called from generated code
+    inline fun <reified T : Any> setList(obj: RealmObjectInternal, col: String, list: RealmList<Any?>) {
+        val existingList = getList<T>(obj, col)
+        if (list !is ManagedRealmList || !RealmInterop.realm_equals(existingList.nativePointer, list.nativePointer)) {
+            existingList.also {
+                it.clear()
+                it.addAll(list)
+            }
+        }
     }
 }