+++ packages/plugin-compiler/src/main/kotlin/io/realm/compiler/AccessorModifierIrGeneration.kt
@@ -354,6 +354,7 @@ class AccessorModifierIrGeneration(private val pluginContext: IrPluginContext) {
                 modifyAccessor(
                     property = declaration,
                     getFunction = getList,
+                    setFunction = setList,
                     collectionType = CollectionType.LIST
                 )
             }